requirements.txt
openai
httpx
pydantic
fastapi
psycopg2
dotenv
langgraph
uvicorn
python-docx
python-multipart